create auth

+++ src/modules/user/services/user.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -8,6 +9,7 @@ import { Like, Repository } from 'typeorm';
 import { UserEntity } from '../entities/user.entity';
 import { CreateUserPayload } from '../models/create-user.payload';
 import { UpdateUserPayload } from '../models/update-user.payload';
+import * as bcryptjs from 'bcryptjs';
 
 @Injectable()
 export class UserService {
@@ -15,6 +17,10 @@ export class UserService {
     @InjectRepository(UserEntity)
     private readonly repository: Repository<UserEntity>,
   ) {}
+
+  public getRepository(): Repository<UserEntity> {
+    return this.repository;
+  }
 
   public async getUsers(search: string): Promise<UserEntity[]> {
     const users = await this.repository.find({
@@ -27,8 +33,8 @@ export class UserService {
     return users;
   }
 
-  public async getOneUser(userId: string): Promise<UserEntity> {
-    const user = await this.repository.findOneBy({ id: +userId });
+  public async getOneUser(userId: number): Promise<UserEntity> {
+    const user = await this.repository.findOneBy({ id: userId });
 
     if (!user) throw new NotFoundException('O usuário não foi encontrado');
 
@@ -45,9 +51,11 @@ export class UserService {
 
     const user = new UserEntity();
 
+    const passwordSalt = await bcryptjs.genSalt();
+
     user.name = payload.name;
     user.email = payload.email;
-    user.password = payload.password;
+    user.password = await bcryptjs.hash(payload.password, passwordSalt);
     user.role = payload.role;
     user.imageUrl = payload.imageUrl;
 
@@ -55,12 +63,18 @@ export class UserService {
   }
 
   public async putUser(
+    requestUser: UserEntity,
     userId: string,
     payload: UpdateUserPayload,
   ): Promise<UserEntity> {
     const user = await this.repository.findOneBy({ id: +userId });
 
     if (!user) throw new NotFoundException('O usuário não foi encontrado');
+
+    if (requestUser.id !== user.id)
+      throw new ForbiddenException(
+        'Você não tem permissão para atualizar esse usuário',
+      );
 
     user.name = payload.name ?? user.name;
     user.role = payload.role ?? user.role;
@@ -69,10 +83,18 @@ export class UserService {
     return await this.repository.save(user);
   }
 
-  public async deleteUser(userId: string): Promise<void> {
+  public async deleteUser(
+    requestUser: UserEntity,
+    userId: string,
+  ): Promise<void> {
     const user = await this.repository.findOneBy({ id: +userId });
 
     if (!user) throw new NotFoundException('O usuário não foi encontrado');
+
+    if (requestUser.id !== user.id)
+      throw new ForbiddenException(
+        'Você não tem permissão para deletar esse usuário',
+      );
 
     await this.repository.remove(user);
   }
